PyTest DQ Framework/requirements.txt
psycopg2-binary
pytest
pyarrow
pandas
pytest-html